spatie install and configured

+++ backend/database/migrations/2024_07_03_144240_roles_user_tb.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('roles_user', function (Blueprint $table) {
             $table->id();
-            $table->integer('role_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('role_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('role_id')->references('id')->on('role_tb');
             $table->foreign('user_id')->references('id')->on('user_tb');
             $table->timestamps();
